requirements.txt
fastapi
uvicorn[standard]
diffusers
transformers
torch
Pillow